Web article tests will now have the correct snapshots and not be based on iOS.
