+++ packages/article/__tests__/android/jest.config.js
@@ -3,6 +3,7 @@ const jestConfigurator = require("@times-components/jest-configurator");
 module.exports = Object.assign(jestConfigurator("article", "android"), {
   collectCoverageFrom: [
     "**/packages/article/*.js",
+    "**/packages/article/**/*.android.js",
     "!**/packages/article/*.web.js",
     "!**/packages/article/*stories*",
     "!**/packages/article/__tests__/**",
