+++ packages/article/__tests__/ios/jest.config.js
@@ -3,7 +3,9 @@ const jestConfigurator = require("@times-components/jest-configurator");
 module.exports = Object.assign(jestConfigurator("article", "ios"), {
   collectCoverageFrom: [
     "**/packages/article/*.js",
-    "!**/packages/article/*.web.js",
+    "**/packages/article/styles/**/*.js",
+    "!**/packages/article/**/*.web.js",
+    "!**/packages/article/**/*.android.js",
     "!**/packages/article/*stories*",
     "!**/packages/article/__tests__/**",
     "!**/packages/article/*fructose*",
