--- packages/article/styles/header/index.ios.js
@@ -1,22 +1,0 @@
-import { StyleSheet } from "react-native";
-import globalStyle from "../shared";
-import sharedStyles from "./shared";
-
-const nativeStyles = {
-  ...sharedStyles,
-  articleHeadLineText: {
-    ...sharedStyles.articleHeadLineText,
-    fontWeight: "700"
-  },
-  standFirst: {
-    ...sharedStyles.standFirst,
-    lineHeight: 25
-  }
-};
-
-const styles = StyleSheet.create({
-  ...globalStyle,
-  ...nativeStyles
-});
-
-export default styles;
