+++ packages/article/styles/header/index.js
@@ -2,9 +2,21 @@ import { StyleSheet } from "react-native";
 import globalStyle from "../shared";
 import sharedStyles from "./shared";
 
+const nativeStyles = {
+  ...sharedStyles,
+  articleHeadLineText: {
+    ...sharedStyles.articleHeadLineText,
+    fontWeight: "700"
+  },
+  standFirst: {
+    ...sharedStyles.standFirst,
+    lineHeight: 25
+  }
+};
+
 const styles = StyleSheet.create({
   ...globalStyle,
-  ...sharedStyles
+  ...nativeStyles
 });
 
 export default styles;
